Refactor layout to one file
- be consistent with most other apps
- prevent circular imports
