+++ src/onegov/agency/layout.py
@@ -1,14 +1,22 @@
 from cached_property import cached_property
-from onegov.agency.collections import ExtendedAgencyCollection
+from onegov.agency.collections import ExtendedAgencyCollection, \
+    ExtendedPersonCollection
 from onegov.agency.models import AgencyMembershipMoveWithinAgency
 from onegov.agency.models import AgencyMove
+from onegov.agency.models.move import AgencyMembershipMoveWithinPerson
 from onegov.core.elements import Confirm
 from onegov.core.elements import Intercooler
 from onegov.core.elements import Link
 from onegov.core.elements import LinkGroup
 from onegov.org import _
-from onegov.org.layout import AdjacencyListLayout
-from onegov.agency.layouts.default import DefaultLayout
+from onegov.org.layout import AdjacencyListLayout, PersonLayout, \
+    PersonCollectionLayout
+
+from onegov.org.layout import DefaultLayout as OrgDefaultLayout
+
+
+class DefaultLayout(OrgDefaultLayout):
+    pass
 
 
 class MoveAgencyMixin(object):
@@ -197,3 +205,97 @@ class AgencyLayout(AdjacencyListLayout, MoveAgencyMixin):
             self.request.link(
                 AgencyMembershipMoveWithinAgency.for_url_template())
         )
+
+
+class AgencyPathMixin(object):
+
+    def agency_path(self, agency):
+        return ' > '.join((
+            self.request.translate(bc.text)
+            for bc in AgencyLayout(agency, self.request).breadcrumbs[2:]
+        ))
+
+
+class MembershipLayout(DefaultLayout):
+
+    @cached_property
+    def breadcrumbs(self):
+        return AgencyLayout(self.model.agency, self.request).breadcrumbs + [
+            Link(self.model.title, self.request.link(self.model))
+        ]
+
+    @cached_property
+    def editbar_links(self):
+        if self.request.is_manager:
+            return [
+                Link(
+                    text=_("Edit"),
+                    url=self.request.link(self.model, 'edit'),
+                    attrs={'class': 'edit-link'}
+                ),
+                Link(
+                    text=_("Delete"),
+                    url=self.csrf_protected_url(
+                        self.request.link(self.model)
+                    ),
+                    attrs={'class': 'delete-link'},
+                    traits=(
+                        Confirm(
+                            _("Do you really want to delete this membership?"),
+                            _("This cannot be undone."),
+                            _("Delete membership"),
+                            _("Cancel")
+                        ),
+                        Intercooler(
+                            request_method='DELETE',
+                            redirect_after=self.request.link(self.model.agency)
+                        )
+                    )
+                )
+            ]
+
+
+class ExtendedPersonCollectionLayout(PersonCollectionLayout, AgencyPathMixin):
+
+    @cached_property
+    def editbar_links(self):
+        if self.request.is_manager:
+            return [
+                Link(
+                    text=_("Create Excel"),
+                    url=self.request.link(
+                        self.model, name='create-people-xlsx'),
+                    attrs={'class': 'create-excel'}
+                ),
+                LinkGroup(
+                    title=_("Add"),
+                    links=[
+                        Link(
+                            text=_("Person"),
+                            url=self.request.link(
+                                self.model,
+                                name='new'
+                            ),
+                            attrs={'class': 'new-person'}
+                        )
+                    ]
+                ),
+            ]
+
+
+class ExtendedPersonLayout(PersonLayout, AgencyPathMixin):
+
+    @cached_property
+    def collection(self):
+        return ExtendedPersonCollection(self.request.session)
+
+    @cached_property
+    def move_membership_within_person_url_template(self):
+        return self.csrf_protected_url(
+            self.request.link(
+                AgencyMembershipMoveWithinPerson.for_url_template())
+        )
+
+
+class AgencySearchLayout(DefaultLayout, AgencyPathMixin):
+    pass
