--- src/onegov/agency/layouts/__init__.py
@@ -1,14 +1,0 @@
-from onegov.agency.layouts.agency import AgencyCollectionLayout
-from onegov.agency.layouts.agency import AgencyLayout
-from onegov.agency.layouts.membership import MembershipLayout
-from onegov.agency.layouts.people import ExtendedPersonCollectionLayout
-from onegov.agency.layouts.people import ExtendedPersonLayout
-
-
-__all__ = (
-    'AgencyCollectionLayout',
-    'AgencyLayout',
-    'ExtendedPersonCollectionLayout',
-    'ExtendedPersonLayout',
-    'MembershipLayout',
-)
