--- src/onegov/agency/layouts/default.py
@@ -1,6 +1,0 @@
-
-from onegov.org.layout import DefaultLayout as OrgDefaultLayout
-
-
-class DefaultLayout(OrgDefaultLayout):
-    pass
