+++ src/onegov/agency/models/ticket.py
@@ -1,8 +1,8 @@
 from cached_property import cached_property
 from onegov.agency.collections import ExtendedAgencyCollection
 from onegov.agency.collections import ExtendedPersonCollection
-from onegov.agency.layouts import AgencyLayout
-from onegov.agency.layouts import ExtendedPersonLayout
+from onegov.agency.layout import AgencyLayout
+from onegov.agency.layout import ExtendedPersonLayout
 from onegov.core.templates import render_macro
 from onegov.org import _
 from onegov.org.models.ticket import OrgTicketMixin
