+++ tests/onegov/agency/test_layouts.py
@@ -1,10 +1,10 @@
 from onegov.agency.collections import ExtendedAgencyCollection
 from onegov.agency.collections import ExtendedPersonCollection
-from onegov.agency.layouts import AgencyCollectionLayout
-from onegov.agency.layouts import AgencyLayout
-from onegov.agency.layouts import ExtendedPersonCollectionLayout
-from onegov.agency.layouts import ExtendedPersonLayout
-from onegov.agency.layouts import MembershipLayout
+from onegov.agency.layout import AgencyCollectionLayout
+from onegov.agency.layout import AgencyLayout
+from onegov.agency.layout import ExtendedPersonCollectionLayout
+from onegov.agency.layout import ExtendedPersonLayout
+from onegov.agency.layout import MembershipLayout
 from onegov.agency.models import ExtendedAgency
 from onegov.agency.models import ExtendedPerson
 from onegov.people import AgencyMembership
